First Wave of Sonar Cloud Refactors for AB#12839
