+++ Apps/WebClient/src/ClientApp/src/components/CommunicationComponent.vue
