+++ Apps/WebClient/src/ClientApp/src/components/ReleaseNoteComponent.vue
