+++ Apps/WebClient/src/ClientApp/src/components/ResourceCentreComponent.vue
