+++ Apps/WebClient/src/ClientApp/src/components/calendar/CalendarBodyComponent.vue
