+++ Apps/WebClient/src/ClientApp/src/components/shared/hgIconComponent.vue
