+++ Apps/WebClient/src/ClientApp/src/components/shared/pageTitleComponent.vue
