+++ Apps/WebClient/src/ClientApp/src/components/timeline/entryCard/covid19LaboratoryOrder.vue
@@ -6,7 +6,7 @@ import Vue from "vue";
 import { Component, Prop, Ref } from "vue-property-decorator";
 import { Getter } from "vuex-class";
 
-import Covid19LaboratoryTestDescriptionComponent from "@/components/laboratory/covid19LaboratoryTestDescription.vue";
+import Covid19LaboratoryTestDescriptionComponent from "@/components/laboratory/covid19LaboratoryTestDescriptionComponent.vue";
 import MessageModalComponent from "@/components/modal/genericMessage.vue";
 import Covid19LaboratoryOrderTimelineEntry from "@/models/covid19LaboratoryOrderTimelineEntry";
 import { DateWrapper } from "@/models/dateWrapper";
