+++ Apps/WebClient/src/ClientApp/src/main.ts
@@ -29,12 +29,12 @@ import VueRouter from "vue-router";
 import Vuelidate from "vuelidate";
 import { Store } from "vuex";
 
-import HgButtonComponent from "@/components/shared/hgButton.vue";
-import HgCardButtonComponent from "@/components/shared/hgCardButton.vue";
-import HgDropdownComponent from "@/components/shared/hgDropdown.vue";
-import HgIconComponent from "@/components/shared/hgIcon.vue";
-import PageTitleComponent from "@/components/shared/pageTitle.vue";
-import StatusLabelComponent from "@/components/shared/statusLabel.vue";
+import HgButtonComponent from "@/components/shared/hgButtonComponent.vue";
+import HgCardButtonComponent from "@/components/shared/hgCardButtonComponent.vue";
+import HgDropdownComponent from "@/components/shared/hgDropdownComponent.vue";
+import HgIconComponent from "@/components/shared/hgIconComponent.vue";
+import PageTitleComponent from "@/components/shared/pageTitleComponent.vue";
+import StatusLabelComponent from "@/components/shared/statusLabelComponent.vue";
 
 import { ExternalConfiguration } from "@/models/configData";
 import User from "@/models/user";
