+++ Apps/WebClient/src/ClientApp/src/views/covid19.vue
@@ -11,10 +11,10 @@ import Vue from "vue";
 import { Component, Ref, Watch } from "vue-property-decorator";
 import { Action, Getter } from "vuex-class";
 
-import LoadingComponent from "@/components/loading.vue";
+import LoadingComponent from "@/components/LoadingComponent.vue";
 import MessageModalComponent from "@/components/modal/genericMessage.vue";
 import BreadcrumbComponent from "@/components/navmenu/breadcrumb.vue";
-import VaccineCardComponent from "@/components/vaccineCard.vue";
+import VaccineCardComponent from "@/components/VaccineCardComponent.vue";
 import { VaccinationState } from "@/constants/vaccinationState";
 import BreadcrumbItem from "@/models/breadcrumbItem";
 import type { WebClientConfiguration } from "@/models/configData";
