+++ Apps/WebClient/src/ClientApp/src/views/home.vue
@@ -18,7 +18,7 @@ import Vue from "vue";
 import { Component, Ref, Watch } from "vue-property-decorator";
 import { Action, Getter } from "vuex-class";
 
-import LoadingComponent from "@/components/loading.vue";
+import LoadingComponent from "@/components/LoadingComponent.vue";
 import AddQuickLinkComponent from "@/components/modal/addQuickLink.vue";
 import MessageModalComponent from "@/components/modal/genericMessage.vue";
 import { EntryType, entryTypeMap } from "@/constants/entryType";
