+++ Apps/WebClient/src/ClientApp/src/views/publicVaccineCard.vue
@@ -10,11 +10,11 @@ import { Validation } from "vuelidate/vuelidate";
 import { Action, Getter } from "vuex-class";
 
 import Image06 from "@/assets/images/landing/006-BCServicesCardLogo.png";
-import ErrorCardComponent from "@/components/errorCard.vue";
-import LoadingComponent from "@/components/loading.vue";
+import ErrorCardComponent from "@/components/ErrorCardComponent.vue";
+import LoadingComponent from "@/components/LoadingComponent.vue";
 import MessageModalComponent from "@/components/modal/genericMessage.vue";
-import HgDateDropdownComponent from "@/components/shared/hgDateDropdown.vue";
-import VaccineCardComponent from "@/components/vaccineCard.vue";
+import HgDateDropdownComponent from "@/components/shared/hgDateDropdownComponent.vue";
+import VaccineCardComponent from "@/components/VaccineCardComponent.vue";
 import { VaccinationState } from "@/constants/vaccinationState";
 import BannerError from "@/models/bannerError";
 import type { WebClientConfiguration } from "@/models/configData";
