+++ Apps/WebClient/src/ClientApp/src/views/timeline.vue
@@ -6,12 +6,12 @@ import Vue from "vue";
 import { Component, Watch } from "vue-property-decorator";
 import { Action, Getter } from "vuex-class";
 
-import ErrorCardComponent from "@/components/errorCard.vue";
+import ErrorCardComponent from "@/components/ErrorCardComponent.vue";
 import CovidTestModalComponent from "@/components/modal/covidTestModal.vue";
 import NoteEditComponent from "@/components/modal/noteEdit.vue";
 import ProtectiveWordComponent from "@/components/modal/protectiveWord.vue";
 import BreadcrumbComponent from "@/components/navmenu/breadcrumb.vue";
-import ResourceCentreComponent from "@/components/resourceCentre.vue";
+import ResourceCentreComponent from "@/components/ResourceCentreComponent.vue";
 import AddNoteButtonComponent from "@/components/timeline/addNoteButton.vue";
 import CalendarTimelineComponent from "@/components/timeline/calendarTimeline.vue";
 import EntryDetailsComponent from "@/components/timeline/entryCard/entryDetails.vue";
